+++ src/user/controllers/create_user.go
@@ -12,12 +12,12 @@ func CreateUser(c *gin.Context) {
 	var user user_dtos.CreateUserRequest
 	if err := c.ShouldBindJSON(&user); err != nil {
 		statusCode := http.StatusUnprocessableEntity
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request body", "statusCode": statusCode})
+		c.JSON(statusCode, gin.H{"error": "invalid request body", "statusCode": statusCode})
 		return
 	}
 	if err := user_dtos.Validate(&user); err != nil {
 		statusCode := http.StatusUnprocessableEntity
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error(), "statusCode": statusCode})
+		c.JSON(statusCode, gin.H{"error": err.Error(), "statusCode": statusCode})
 		return
 	}
 	create_user_usecase.Createuser(c, user)