+++ documentation/changelog.md
@@ -1,6 +1,8 @@
 # [v0.0.1] - XX/XX/XXXX
 
+- create_user use case, all tests and documentation
+- Setup integration tests with testcontainers (https://testcontainers.com/), up postgres container for tests, run migrations and clean database before each test
 - CI/CD to exec lint, build and automated tests
 - Enable all possible security features on the GitHub platform
 - Fisrt version of code documentation
-- Create first migrations
+- Migration to create user table